migrated out of edge based graph factory
